% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayBoundaryAverage.R
\name{arrayBoundaryAverage}
\alias{arrayBoundaryAverage}
\title{Averages the boundary elements of an array}
\usage{
arrayBoundaryAverage(pj1pj2Matrix, var2 = FALSE)
}
\arguments{
\item{pj1pj2Matrix}{Empirical copula matrix. It is assumed that the periodic variable is the first column of the argument \code{uv} in the function \code{\link{pj1j2Matrix}}.}

\item{var2}{Unimplemented yet. Which is the periodic random variable? Default to the first column of the argument \code{uv} in the function \code{\link{pj1j2Matrix}}.}
}
\value{
A numeric array with equall (averaged) boundary elements.
}
\description{
This function was implemented in order to reproduce the average of page 1998 of \href{https://link.springer.com/article/10.1007/s00477-013-0733-y}{Carnicero et al., 2013}.
}
\details{
This function was implemented in order to reproduce the average of page 1998 of \href{https://link.springer.com/article/10.1007/s00477-013-0733-y}{Carnicero et al., 2013}. By default the periodic variable is the one of the first colum of the argument \code{uv} of the function pj1j2Matrix.
}
\examples{
ne <- 30
set.seed(12)
uve <- cbind(runif(ne), runif(ne))
library(squash)
arrayBoundaryAverage(pj1j2Matrix(uv = uve, k = 3)$counts)
}
\author{
Francisco Mendoza-Torres (\email{mentofran@gmail.com})
}
