% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependenceAnalysisPlot.R
\name{dependenceAnalysisPlot}
\alias{dependenceAnalysisPlot}
\title{Bivariate dependence analysis plot}
\usage{
dependenceAnalysisPlot(x = NULL, asp = 1, main = "", xGridBreaks = c(0,
  0.25, 0.5, 0.75, 1), yGridBreaks = xGridBreaks, xlab = "x", ylab = "y",
  xlabels = TRUE, ylabels = TRUE, xlSide = 1, ylSide = 2,
  col = par("col"), colp = c("white", "gray90"), grid = NULL, ...)
}
\arguments{
\item{x}{table (2-colum matrix or data.frame) of, possibly pseudo-, observations}

\item{asp}{The \code{y/x} aspect ratio, see \link[graphics]{plot.window}.}

\item{main}{Character. Main title for the plot. See \link[graphics]{title}.}

\item{xGridBreaks, yGridBreaks}{zone delimiters. Default to Inter-Quartile Area (iqa := \code{c(0, .25, .5, .75, 1) x c(0, .25, .5, .75, 1)}).}

\item{xlab, ylab}{Character. Label of the axis.}

\item{xlabels, ylabels}{Character. Labels of the annotations of the axis ticks marks.}

\item{xlSide, ylSide}{Side where the labels of the axis must be plotted. See \code{side} parameter of \link[graphics]{axis}.}

\item{col}{color of the points.}

\item{colp}{color of the pixels.}

\item{grid}{list. grid constructed with (posibly pseudo-) observations of modes, minima,
inflexion points or any other points of interest of both variables.
 \code{grid} is passed to \link[graphics]{abline}.}

\item{...}{further parameters to \link[graphics]{points.default}.}
}
\description{
Plot scatterplots and pseudo-observations plots over a rectangular domain partition by rectangular areas (also called pixes). This function can be used with the function \link{plottingPosition} for the pseudo-observations.
}
\examples{
# Example 1: pseudo-observations
ne <- 30
ue <- matrix(runif(ne), ne/2, 2)
xpixelse <- c(0, .25, .5, .75, 1)
dependenceAnalysisPlot(x = ue, xGridBreaks = xpixelse) # colp = rainbow(n = npixels),

# Example 2: observations
ne <- 100
ue <- matrix(rnorm(ne), ne/2, 2)
xpixelse <- quantile(ue[, 1])
ypixelse <- quantile(ue[, 2])
dependenceAnalysisPlot(x = ue,  asp = NA,
    xlabels = c("min", "q1", "M", "q3", "max"),
    xGridBreaks = xpixelse,
    yGridBreaks = ypixelse)

npixels <- (length(xpixelse) - 1) * (length(ypixelse) - 1)
dependenceAnalysisPlot(x = ue,
    xGridBreaks = xpixelse,
    yGridBreaks = ypixelse,
    colp = rainbow(n = npixels),
    pch = 20, cex = 2, asp = NA,
    grid = list(h = 0, v = c(-1, 1),
                lty = 2, col = "blue"))
}
\author{
Francisco Mendoza-Torres (\email{mentofran@gmail.com})
}
