% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalCDF2Dcounts.R
\name{empiricalCDF2Dcounts}
\alias{empiricalCDF2Dcounts}
\title{Computes the empirical bivariate distribution and store the results in a matrix}
\usage{
empiricalCDF2Dcounts(x)
}
\arguments{
\item{x}{A 2-columns data.frame with numeric observations \eqn{(x1,y1),...,(xn,yn)}.}
}
\value{
An integer \eqn{(n+1)x(n+1)}-matrix with the values of the empirical copula multiplied by $n$.
}
\description{
Computes the matrix of counts for the empirical distribution.
To get the real empirical distribution values divede the result by $n$.
This function considers that there are no duplicated values both in $x$ and in $y$.
}
\details{
\code{x[i+1,j+1]} is equivalent to \eqn{Cn(i/n,j/n)}. Deheuvels, 1979 or Nelsen, 2006.
}
\examples{
# Example 1
library(squash)
exy <- cbind(1:5, c(2, 4, 3, 6, 7)); print(exy)
set.seed(1); exy <- exy[sample(1:5), ]
print(empiricalCDF2Dcounts(exy))
# to compute the empirical distribution values divide by n
(1/5)* empiricalCDF2Dcounts(exy)

# Example 2
x <- cbind(1:5, c(4, 2, 1, 5, 3))
print(empiricalCDF2Dcounts(x))
# to compute the empirical distribution values divide by n
(1/5)* empiricalCDF2Dcounts(x)

# Example 3 (Error catch):
exy <- cbind(x=c(1:3,2),y=c(1,2,3,2))
# empiricalCDF2Dcounts(exy)
}
\references{
Equation (3) of \href{https://link.springer.com/chapter/10.1007/978-3-642-12465-5_13}{Erdely, A. and Diaz-Viera, M. 2010}.
}
\author{
Francisco Mendoza-Torres (\email{mentofran@gmail.com}) and Arturo Erdely (\email{arturo.erdely@comunidad.unam.mx})
}
