% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalDistribution.R
\docType{package}
\name{empiricalDistribution}
\alias{empiricalDistribution}
\alias{empiricalDistribution-package}
\title{Empirical distributions}
\description{
Empirical distribution computation
}
