% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmat.copem.R
\name{genmat.copem}
\alias{genmat.copem}
\title{Matrix of empirical copula.}
\usage{
genmat.copem(mat.xy)
}
\arguments{
\item{mat.xy}{A 2-columns data.frame with the observations \eqn{(x1,y1),...,(xn,yn)}.}
}
\value{
A \eqn{(n+1)x(n+1)}- matrix with the values of the empirical copula.
}
\description{
Computes the matrix of empirical copula.
}
\details{
\code{mat.copem[i+1,j+1]} is equivalent to \eqn{Cn(i/n,j/n)}. Deheuvels, 1979 or Nelsen, 2006
}
\examples{
x <- cbind(1:5, c(2, 4, 3, 6, 7))
genmat.copem(x)
}
\references{
Equation (3) of \href{https://link.springer.com/chapter/10.1007/978-3-642-12465-5_13}{Erdely, A. and Diaz-Viera, M. 2010}.
}
