% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2D.R
\name{hist2D}
\alias{hist2D}
\title{2D histogram.}
\usage{
hist2D(x, y = NULL, nx = 50, ny = nx, xlim = NULL, ylim = NULL,
  xbreaks = NULL, ybreaks = NULL, plot = TRUE, xlab = NULL,
  ylab = NULL, zlab = "Counts", colFn = heat, breaks = prettyInt,
  right = TRUE, ...)
}
\arguments{
\item{x}{ Numeric vectors. }

\item{y}{same as x}

\item{nx}{ Approximate number of intervals along x and y axes. }

\item{ny}{same as nx}

\item{xlim}{ Limit the range of data points considered. }

\item{ylim}{same as xlim}

\item{xbreaks}{ Breakpoints between bins along x and y axes. }

\item{ybreaks}{same as xbreaks}

\item{plot}{ Plot the histogram? }

\item{xlab}{ Axis labels. }

\item{ylab}{same as ylab}

\item{zlab}{ Label for the color key. }

\item{colFn}{ Color key parameters; see \code{\link{makecmap}}. }

\item{breaks}{see \link[graphics]{hist}}

\item{right}{see \link[graphics]{hist}}

\item{...}{ Further arguments passed to \code{\link{colorgram}}. }
}
\description{
Creates a 2D version of a histogram.
This function is used by the function "pj1pj2Matrix" found in the file pj1pj2 for directinal data.
}
\examples{
ne <- 20
set.seed(12)
uve <- cbind(runif(ne), runif(ne))
print(hist2D(uve, nx = 3))
}
\author{
Francisco Mendoza-Torres (\email{mentofran@gmail.com})
}
