% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj1j2Matrix.R
\name{pj1j2Matrix}
\alias{pj1j2Matrix}
\title{Empirical counts/density computing}
\usage{
pj1j2Matrix(uv, k = NULL, plot = FALSE, ...)
}
\arguments{
\item{uv}{2-columns data.frame of nonexedance probabilities.}

\item{k}{A positive integer. the order of the Bernstein polynomial.
Default to cubic root of the number of observations as recommended in page 1998, Carnicero, 2013.}

\item{plot}{Plot pixels form of 2D histogram? Default to FALSE}

\item{...}{Other arguments passed to \code{hist2D}.}
}
\value{
A 2D matrix whose elements (pj1pj2) are indexed the same way as pj1j2 in formula 5 of Carnicero.
}
\description{
Computes formula (5) of \href{https://link.springer.com/article/10.1007/s00477-013-0733-y}{Carnicero et al., 2013}.
}
\details{
Here comes the details.
}
\examples{
ne <- 30
set.seed(12)
uve <- cbind(runif(ne), runif(ne))
pj1j2Matrix(uv = uve, k = 3, plot = TRUE)
}
\references{
\href{https://link.springer.com/article/10.1007/s00477-013-0733-y}{Carnicero et al., 2013}
}
\author{
Francisco Mendoza-Torres (\email{mentofran@gmail.com})
}
