% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProbs.R
\name{plotProbs}
\alias{plotProbs}
\title{Plot pseudo-observations}
\usage{
plotProbs(..., bty = "o", xlim = 0:1, ylim = 0:1, atx = c(0, 0.5, 1),
  aty = atx)
}
\arguments{
\item{...}{Arguments passed to \link[graphics]{plot.default}}

\item{bty}{Bounding box type. See \link[graphics]{par}. A character string which determined the type of box which is drawn about plots. If bty is one of "o" (the default), "l", "7", "c", "u", or "]" the resulting box resembles the corresponding upper case letter. A value of "n" suppresses the box.}

\item{xlim}{the x limits (x1, x2) of the plot.  Note that \code{x1 > x2}
    is allowed and leads to a \sQuote{reversed axis}.

    The default value, \code{NULL}, indicates that the range of the
    \link{finite} values to be plotted should be used.}

\item{ylim}{the y limits of the plot.}

\item{atx, aty}{The points at which tick-marks are to be drawn. See \code{at} of \link[graphics]{axis}.}
}
\description{
Scatterplot of pseudo-observations (values in \code{[0,1]}) that preserves the aspect ratio (asp=1).
}
\examples{
uv <- cbind(runif(100), runif(100))
plotProbs(uv)
}
\author{
Francisco Mendoza-Torres (\email{mentofran@gmail.com})
}
